% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Produce a graph for desired country of year vs population}
\usage{
CountryPopulation(country)
}
\arguments{
\item{a}{A valid country in string format 'country'}
}
\value{
ggplot of desired country x = year y = population
}
\description{
Produce a graph for desired country of year vs population
}
