% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup attendance by total attendance and average attendance per event}
\format{
a data frame with 22 observations of 4 variables
\describe{
  \item{Totalattendance}{total attendance of world cup}
  \item{Matches}{number of matches at world cup}
  \item{Averageattendance}{average attendance per match}
  \item{WorldCup}{World cup by year and host country}
}
}
\usage{
World_Cup
}
\description{
World Cup attendance by total attendance and average attendance per event
}
\keyword{datasets}
