% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World population by country from 1950 to 2020}
\format{
a data frame with 263 observations of 72 variables
\describe{
  \item{CountryName}{Name of country ordered alphabetically by region}
  \item{Year}{Year of population statistic}
}
}
\usage{
WorldPopulation
}
\description{
World population by country from 1950 to 2020
}
\keyword{datasets}
